Ephemeral room support (#367)
# Description

This adds support for Ephemeral rooms.

Ephemeral rooms are instances of rooms that are unique and get their own RoomId's. They cannot be saved to disk, and are cleaned up when all players leave the group of instanced/ephemeral rooms.

This is accomplished by allocating a range of RoomId's that are used only by ephemeral ID's. Room's being copied get their RoomId replaced with a newly generated one.

## Changes

- Eliminated Tutorial Room copies so there is now only one of each tutorial room.
- Starting the tutorial now creates an ephemeral copy and puts the player in it.
- Fixed up tutorial scripts to add a little more delay in tutorial instructions.
- Updated config.yaml so that `TutorialStartRooms` is now a complete list of all tutorial rooms, starting with the "tutorial start" room.
- `rooms/ephemeral.go` handles ephemeral room specifics.
- Added some script cleanup to scripting, various event handlers, etc. all for managing the pieces of this.
- Cleaned up some code and moved to its own event handlers, such as the Guide spawning checks.