avoiding overflow warning for 32/64 bit system (#370)
# Description

Building for 32-bit is giving this error:
`cannot use roomIdMin64 (untyped int constant 1000000000000) as int value in assignment (overflows)`

Even though there is runtime detection for the integer bitsize, we still need to trick the compiler.

This sets it to the 32 bit value and then multiplies it by 1000 if the maxint is over this number.

## Changes

- Changed the way we check for which minimum ephemeral RoomId value to use.