Mudlet GMCP Discord Rich Presence functionality (#374)
# Description

This adds the ability for the game to send Rich Presence information to Discord.

<img width="290" alt="Screenshot 2025-05-10 at 12 46 06" src="https://github.com/user-attachments/assets/412a6d97-3517-421e-b5b8-86e09fb90024" />

This will also add a better Discord integration directly in Mudlet:

<img width="529" alt="Screenshot 2025-05-10 at 12 50 34" src="https://github.com/user-attachments/assets/ee1e4535-fd88-4dc2-903e-78d494a1361d" />

By default it sends the following information:
```
DiscordApplicationID: "1298377884154724412",           // Default GoMud Discord Server
DiscordInviteURL:     "https://discord.gg/FaauSYej3n", // Default GoMud Discord Server
DiscordDetails:       "Using GoMudEngine",
DiscordState:         "Exploring the world",
DiscordLargeImageKey: "server-icon",                   // GoMud Discord Icon
DiscordSmallImageKey: "character-icon",                // Hover icon (not used)
```
As well as player name, level, area, party information and connection time. This information can be change in the module config.yaml

The game name used in Discord comes from the application id or if not sent, the game name in config.yaml

Added "discord" user command to enable or disable certain information:
```
  discord               - Shows current settings and available commands
  
  discord info on|off   - Enable/disable Discord application info
  discord status on|off - Enable/disable Discord status updates

  discord area on|off   - Show/hide current area in status
  discord party on|off  - Show/hide party information
  discord name on|off   - Show/hide character name
  discord level on|off  - Show/hide character level
```
Everything is saved on the player, and kept across sessions.

Added helpfile in the Integration category. Here we have potential for other integrations to be placed.

This gmcp information could potentially be used by the web-client as well, to send information to Discord if activated.

I snuck in some beautification and alignment on the other helpfiles for gmcp and Mudlet - that is why they are also in this PR.