Item purchase scripting (#412)
# Description

This PR primarily adds an `onPurchase()` script event for items.

The intention is that an admin can create an item listed in a shop that can be scripted to do creative things that we may not anticipate, allowing them to leverage shops for other purchases.

An initial example has been created/added, which is buying a `room for rent` at the inn.

It also addresses some other small typos and bugs.

## Changes

- Added `onPurchase()` item script event and documentation
- Fixed a typo for a room: `paventment` => `pavement`
- Added an example `onPurchase()` event to the inn shop (`room rental`).
- Fixed `cancel-on-action` buffs to cancel on leaving to an exit
- Renamed "leave" to "downstairs" in Inn room, since "l" (for look) was a little confusingly taking the "leave" exit after sleeping.
- Inn room now uses an ephemeral copy per player, so players don't all end up in one instance of the room.
- roominfo template had an error, referring to old ZoneInfo. This is fixed.
- `buy` was not doing anything when you weren't in a shop. Now provides feedback to the user.
- Some small language updates around the `buy` command feedback.

### Links
https://github.com/GoMudEngine/GoMud/issues/411